STY: use common methods for simulators
